serwis-app/backend/app/auth.py
from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.orm import Session
from fastapi.security import OAuth2PasswordRequestForm
from jose import jwt
from datetime import datetime, timedelta
from app import models, schemas, crud, database, config

router = APIRouter()

@router.post("/login", response_model=schemas.Token)
def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(database.get_db)):
    user = crud.get_admin(db)
    if not user or form_data.username != user.username or not crud.verify_password(form_data.password, user.hashed_password):
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Nieprawidłowy login lub hasło")
    access_token_expires = timedelta(minutes=config.ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = jwt.encode({
        "sub": user.username,
        "exp": datetime.utcnow() + access_token_expires
    }, config.SECRET_KEY, algorithm=config.ALGORITHM)
    return {"access_token": access_token, "token_type": "bearer"} 