serwis-app/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
email-validator
reportlab 